requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
streamlit==1.28.0
langchain==0.0.335
langchain-community==0.0.10
langchain-openai==0.0.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
chromadb==0.4.18
pymupdf==1.23.5
pdfplumber==0.10.2
python-multipart==0.0.6
rank-bm25==0.2.2
ollama==0.1.5
python-dotenv==1.0.0
numpy==1.26.1
pandas==2.1.2
tqdm==4.66.1 